Improved usability for variable legend
When a user would click on a `†i9` variable legend link, the page would jump to the legend. However the legend is contained within a (closed) accordion section, so the user would still not see the legend.

With this fix, the relevant accordion section is auto-expanded when a variable legend link is clicked.